+++ include/TestArray.hpp
@@ -14,27 +14,27 @@ class TestArray{
 private:
   DebugLevel debugLevel_ = silent;
 
-  uint32_t *array_ = nullptr;
-  uint32_t len_ = 100;
+  uint64_t *array_ = nullptr;
+  uint64_t len_ = 100;
 
   std::random_device rd_;
   std::mt19937 rng_;
 
   void generateArray();
-  void convertCharArrayToIntArray(uint32_t len, char *charArray[], uint32_t *intArray);
+  void convertCharArrayToIntArray(uint64_t len, char *charArray[], uint64_t *intArray);
 
 public:
   TestArray();
-  TestArray(uint32_t len);
-  TestArray(uint32_t len, char *array[]);
-  TestArray(std::vector<std::uint32_t> values);
+  TestArray(uint64_t len);
+  TestArray(uint64_t len, char *array[]);
+  TestArray(std::vector<std::uint64_t> values);
 
   ~TestArray();
 
   void setDebugLevel(DebugLevel debugLevel);
 
-  const uint32_t *getArray();
-  uint32_t len() {return len_;};
+  const uint64_t *getArray();
+  uint64_t len() {return len_;};
 
   void printArray();
 };
